test: Input 스토리북 생성
